pyproject.toml
[tool.poetry]
name = "zoneroles-bot"
version = "0.1.0"
description = ""
authors = ["Irshad PI"]

[tool.poetry.dependencies]
python = "^3.10"
discord-py-interactions = "^4.2.0"
aioschedule = "^0.5.2"
python-dotenv = "^0.20.0"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
